catch a convergence exception
